Improve kmer2node counting.
For SNPs we expect to see many node=0 kmer hits.  If previous node num
is X and next node num is X and we have a run of 0 inbetween, then fix
the last KMER-1 entries as we've anchored at least the end of it as a
unique kmer.

Essentially this is counting the number of bases covered by a unique
kmer, so it rescues rare SNPs but won't rescue too large a divergence
with many SNPs in the region.