Adds gaf2seq
This is a tool that combines a GFA file (1st arg) with a GAF alignment
path (2nd arg or stdin) to construct the graph sequence.  This should
be the sequence which our query gets aligned against for purpose of
producing a CIGAR string and NM tags.

It can also be used as a quick check for whether the alignment appears
to be valid.