Correct for maximum kmer count.
We have a notion of the expected kmer hit rate and report according to
how we compare vs expectation.  However if we align in one region (say
100 to 200) and we expect 50 matches but we get 75, then that's not
1.5x coverage if we still only generated hits in a for that node and
nowhere inbetween.  The evidence implies we simply had a higher than
expected hit rate.

So we cap the maximum hit rate at the start..end positions for
nodes in consecutive runs (permitting misses and dups and end
effects).