Correct expected counts for first node and last node.
These may be short if the sequence starts part way.  We could track
very first and very last hit and amend expected, but instead we use
MAX(ratio, hit/(hit+poss+1)) and let the possible-hit calculations
work for us instead which mostly amounts to the same thing bar the
small change from 1st to 2nd or penultimate to last.