Add train8.fa.
This is from an exploration into the minimal set of GRCh38 HLA
sequences that we can use for training the nodeseq file.  It's just 8
sequences but jointly they align against every node.

Obviously we can get better matches with more detailed data, such as
all GRCh38 sequences, but this is sufficient for decent kmer counting.