Add a basic synthetic genome creation tool.
This is very basic.  It starts with a random sequence and then adds
layers of repeats (eg SINEs and LINES), STRs, copy-number variations
and translocated sequence.

We can then replicate and shred the genome consensus and assemble it
to produce a graph.  As we know the original consensus we can validate
potential paths, although there may be several so this it is tricky to
evaluate.  (TODO)