Tweak defaults and add "clumpiness".
This is the notion that the distribution of STRs and CNVs isn't
random, with every new repeat element having a chance to be in the
same proximity as the previous, giving the effect of genomic
hotspots.  Also we have a chance of translocations while iterating
through the copy numbers, so we may get half of an STR at one site and
then the other half at a new location.