Add align_reg script plus other optimisations.
Align_reg.sh can fetch a public Revio HG002 data set and perform a
region based alignment to the PGRC grch38 published pangenome. This is
the first time we're using a proper pangenome instead of an assembler,
and it's showing weaknesses (such as pathfinder not liking zero depth
nodes).   Example usage:

    region=chr6:32578775-32589848; ~/work/quantum/align_reg.sh https://downloads.pacbcloud.com/public/revio/2022Q4/HG002-rep1/analysis/HG002.m84011_220902_175841_s1.GRCh38.bam $region hprc-v1.0-minigraph-grch38.chr6.gfa

Candidate stats separates breaks from contigs.  As an incorrect join
is worse than leaving a pair of contigs as we don't know. (Ie don't
just guess!)

Optimise parameters on miniasm and syncasm.

Use an installed copy of mdbg.