Add split_gfa_nodes.pl.
This sets a maximum length of node and splits long nodes into chains.
This means the path we return will be no more than max-len extended
left / right.

TODO:  "L node-a + node-b - 0M" means node-a links to the end of node-b,
so if node-b is split it should be the last split and not the first
one.