Add gaf2nodeseq.pl.
This is similar to gfa2nodeseq, which takes GFA + GraphAligner GAF
data to produce a series of sequences per node based on the node
sequence itself + aligned data against it.  However it is tailored
towards the GFA alone.

To improve on this (and compensate for the lack of evidential data) it
follows the graph node and prepends sequences by their previous node
sequence so the kmer analysis can still work when given short nodes.
This helps rescue complex paths with a mix of short node data.