Account for possible hits in kmer2node2 ratio scoring.
We don't know where these hits are.  They're just absent due to dups,
but we take a very rough stab (scalable by repeating the "-U" option)
so we give pathfinder a fighting chance of using the repetitive nodes
to concoct a plausible explanation.  This is comparable to the code
there that sets the minimum depth to 1, but more kmer depth aware.