Improve(?) kmer2node2 -U heuristic.
FIXME: we need to do something with location where the kmers are and
the average number of duplicate locations.

Unique kmers at start and end of a node and a whole raft of ambiguous
ones in the middle of the node probably need to be allocated to that
node.  (We'd need a way of similarly modifying the expected score too
so depth ratio is correct).

Kmers which aren't uniquely flanked represent genuine repeats.  We can
drop them, as we do now, or add distribute them randomly.  What works
depends on how the pathfinder copes.