Add a new kmer2node3 tool.
This is a derivative of kmer2node2 with better estimation of kmer
counting where nodes track from one node to another or from one node
to itself but passing through a set of unknown nodes.

TODO:  If the graph has nodes A->B->C and we see A to C kmers present
but B can never be found as it is 100% duplicated, then we could still
consider incrementing B as the required path.  (If there is a direct
A->C transition also then obviously that also becomes a possible
route.)  Tricky and it requires graph knowledge which this tool
doesn't have currently.

Ideally it should be done by the pathfinder, which does have graph
data, but apparently isn't reliably.