Improve GFA generation
We now use minigraph -n 5,30.  The default is 5,5.  This boosts the
number of minimisers needed when mapping, which causes more divergent
data with poorer matches to not create new graph nodes.  This has the
consequence of reducing the node count.

Without this I noted picking 30 of 50 population was giving better
results than all 50 of 50, despite the latter now meaning the training
set and test sets overlap.  With this change it's now very comparable
and we don't suffer a penalty from graph size increases.