Make genome_create more realistic.
Previously we generated N genomes and then chose the first so many for
training and the last ones for testing.  However each genome is
created in order, with later ones being descended from any earlier
one.

This is wrong for two reasons:
1. In a population our ancestors are recent, not potentially ancient!
2. We used old genomes for GFA construction, but in reality we'd be
   subsampling the current population (the leaf nodes).

We now have an additional parameter to generate N genomes with G
generations.

We also shuffle these and pick a random set for training and another
non-overlapping random set for evaluation.

This also needing tuning of various parameters in both genome_create
and minigraph mapping to get good results.