Move the kmer2node specific bits to run_sim_add_gfa_weights_km.sh
Also simplified this making the 3 kmer phase a loop and permitting
adjusting which kmers to use in the config file.

We no longer create 3 separate nodeseq files and instead use the -K
parameter to specify the length of the prefix.  (This works provided
the first kmer specified is the longest.)