Implement a "--subgraph D W" option.
TODO: move to a separate script so the old behaviour doesn't change.
This is important for testing and validation with the paper.

We can use e.g. --subgraph 999 0 as a way of getting new behaviour with
no subgraph production (as the main code does now), but without the
need to change run_sim_solver_qubo.sh default behaviour.  This also permits
us to explore newer copy-number formulae, as done in --pathfinder_graph.